initiate lib
